DSImplicit: The code works with the 145 bus test case but there is some issue with the data in the case file due to which the initial residual is not nearly zero.
The buses having the most initial mismatch are buses 146 and 142.
Looking at the residuals of these buses, it looks like either the bus voltages in the file are incorrect or there is
some line missing between the two. Another point to note is that this case is adapted from WU test case archive that has
145 buses. An extra bus is added in this test case and this extra bus is connected to 10 other buses. Don't know where this
came from?