Generator should return the field current when asked by the exciter
Currently, the generator sets the field current for the exciter when computing
its own residual. This is wrong. The exciter should querry the generator to get
the field current since the field current is an attribute of the generator. To
enable this, (i) made the exciter to hold a pointer to the generator object, (ii)
removed the setFieldCurrent() method for the exciter, (iii) added getFieldCurrent()
methods for the generator.