Remove manual trigger from container build job
This always adds the job to the pipeline and blocks - ignoring the intent of it being an optional dependency downstream. I suppose if we want to rebuild the container, you could always go back to the last pipeline where it ran and rerun. This should be the last place that any files relevant to the container build changed.